Moved server.js and created api folder
