+++ Neighborhood-Safety-App/api/server.js
@@ -1,0 +1,38 @@
+const express = require('express');
+const mongoose = require('mongoose');
+const bodyParser = require('body-parser');
+const cors = require('cors');
+
+const app = express();
+app.use(bodyParser.json());
+app.use(cors({
+  origin: ['http://localhost:3000', 'https://neighborhood-safety-app-nine.vercel.app'],
+}));
+
+const mongoURI = process.env.MONGO_URI;
+mongoose.connect(mongoURI, { useNewUrlParser: true, useUnifiedTopology: true })
+  .then(() => console.log('MongoDB connected'))
+  .catch((err) => console.error('MongoDB connection error:', err));
+
+const reportSchema = new mongoose.Schema({
+  report_id: { type: String, required: true },
+  issueType: String,
+  location: String,
+  description: String,
+  photoUri: String,
+  createdAt: { type: Date, default: Date.now },
+});
+
+const Report = mongoose.model('Report', reportSchema, 'reports');
+
+app.post('/api/reports', async (req, res) => {
+  try {
+    const report = new Report(req.body);
+    await report.save();
+    res.status(201).json({ message: 'Report submitted successfully', report });
+  } catch (err) {
+    res.status(500).json({ error: 'Failed to submit report' });
+  }
+});
+
+module.exports = app;
