--- Neighborhood-Safety-App/server.js
@@ -1,64 +1,0 @@
-require('dotenv').config(); // Load environment variables
-const express = require('express');
-const mongoose = require('mongoose');
-const bodyParser = require('body-parser');
-const cors = require('cors');
-
-// Initialize Express app
-const app = express();
-app.use(bodyParser.json());
-app.use(cors({
-  origin: ['http://localhost:3000', 'https://neighborhood-safety-app-nine.vercel.app','https://neighborhood-safety-app-nine.vercel.app/api/reports'], // Replace with your frontend domains
-}));
-
-// Connect to MongoDB
-const mongoURI = process.env.MONGO_URI; // Use environment variable
-mongoose.connect(mongoURI, { useNewUrlParser: true, useUnifiedTopology: true })
-  .then(() => console.log('MongoDB connected'))
-  .catch((err) => console.error('MongoDB connection error:', err));
-
-// Define a schema for reports
-const reportSchema = new mongoose.Schema({
-  report_id: { type: String, required: true }, // Add the unique report ID
-  issueType: String,
-  location: String,
-  description: String,
-  photoUri: String,
-  createdAt: { type: Date, default: Date.now },
-});
-
-// Create a model for reports (explicitly set collection name to "Reports")
-const Report = mongoose.model('Report', reportSchema, 'reports');
-
-// API endpoint to submit a report
-app.post('/api/reports', async (req, res) => {
-  try {
-    const report = new Report(req.body);
-    await report.save();
-    res.status(201).json({ message: 'Report submitted successfully', report });
-  } catch (err) {
-    res.status(500).json({ error: 'Failed to submit report' });
-  }
-});
-
-// API endpoint to fetch all reports
-app.get('/api/reports', async (req, res) => {
-  try {
-    const reports = await Report.find();
-    res.status(200).json(reports);
-  } catch (err) {
-    res.status(500).json({ error: 'Failed to fetch reports' });
-  }
-});
-
-// Health check endpoint
-app.get('/health', (req, res) => {
-  res.status(200).json({ status: 'OK', message: 'Server is running' });
-});
-
-// Start the server
-const PORT = process.env.PORT || 5000;
-app.listen(PORT, () => console.log(`Server running on port ${PORT}`));
-
-// Export the app for Vercel
-module.exports = app;
