+++ Neighborhood-Safety-App/vercel.json
@@ -5,14 +5,15 @@
     "cleanUrls": true,
     "framework": null,
     "rewrites": [
-    {
-      "source": "/api/:path*",
-      "destination": "/api/:path*"
-    },
-    {
-      "source": "/:path*",
-      "destination": "/"
-    }
-  ]
+      {
+        "source": "/api/:path*",
+        "destination": "/api/:path*"
+      },
+      {
+        "source": "/:path*",
+        "destination": "/"
+      }
+    ]
+    
   }
